Add support for 1 API & misc. bug fixes

+++ TMdbEasy/TmdbObjects/Images/Poster.cs
@@ -12,7 +12,7 @@ namespace TMdbEasy.TmdbObjects.Images
         public string File_path { get; set; }
         public int Height { get; set; }
         public string Iso_639_1 { get; set; }
-        public int Vote_average { get; set; }
+        public double Vote_average { get; set; }
         public int Vote_count { get; set; }
         public int Width { get; set; }
     }
